CONTROLE INTERNO

+++ app/Http/Controllers/CadastrarEstoqueController.php
@@ -15,11 +15,16 @@ use App\Http\Controllers\View;
 
 use DB;
 
+//CONTROLLER COM TODOS OS MÉTODOS ENVOLVENDO O ESTOQUE
 class CadastrarEstoqueController extends Controller
 {
  
     public function index()
     {
+        if(!\Gate::allows('isAdmin')){
+            abort(403,"Não pode executar essa ação");
+        }
+        //ACESSO PRINCIPAL A TELA DE CADASTRO DE ESTOQUE, MOSTRANDO TODOS OS ITENS, LOCALIDADES E FORNECEDORES. 
         $estoque = new Estoque();
         $estoque = $estoque::all();
 
@@ -31,8 +36,6 @@ class CadastrarEstoqueController extends Controller
 
         $fornecedor = new Fornecedor();
         $fornecedor = $fornecedor::all();
-
-        // return view('cadastrar_estoque', ['estoque'=>$estoque], ['fornecedor'=>$fornecedor], ['item'=>$item]);
 
         return view('cadastrar_estoque')
 
@@ -54,27 +57,44 @@ class CadastrarEstoqueController extends Controller
    
     public function store(Request $request)
     {
+
+        if(!\Gate::allows('isAdmin')){
+            abort(403,"Não pode executar essa ação");
+        }
         
-        //Função pra 
+        //CHAMANDO TODAS AS MODELS QUE SERÃO INSERIDAS NA ENTRADA DO ESTOQUE
         $estoque = new Estoque();
         $item = new Item();
         $fornecedor = new Fornecedor();
         $localidade = new Localidade();
 
-        
-        $estoque->quantidade_total = $request->input('quantidade_total') ;
-        $estoque->numero_nf = $request->input('numero_nf') ;
-        $estoque->data_nf = $request->input('data_nf') ;
-        $estoque->data_garantia = $request->input('data_garantia') ;
-        $estoque->idItem = $request->input('idItem') ;
-        $estoque->idFornecedor = $request->input('idFornecedor') ;
-        $estoque->idLocalidade = $request->input('idLocalidade') ;
-        $estoque->data_entrada = now();
+        //VALIDAÇÃO DOS CAMPOS A SEREM PREENCHIDOS 
+        $estoque = $request->validate([
+            'quantidade_total' => 'required',
+            'idItem' => 'required',
+            'idFornecedor' => 'required',
+            'idLocalidade' => 'required',
+        ]);
 
-        $estoque->save();
+        //INSTANCIANDO NOVAMENTE A MODEL "ESTOQUE" PARA FAZER O INSERT NO BANCO
+        $estoque_cadastrar = new Estoque();
 
- 
-        return redirect()->route('home')
+        //INSERINDO OS DADOS DIGITADOS NO INPUT DO FORMULÁRIO DE CADASTRO DE ESTOQUE
+        $estoque_cadastrar->quantidade_total = $request->input('quantidade_total') ;
+        $estoque_cadastrar->numero_nf = $request->input('numero_nf') ;
+        $estoque_cadastrar->data_nf = $request->input('data_nf') ;
+        $estoque_cadastrar->data_garantia = $request->input('data_garantia') ;
+        $estoque_cadastrar->idItem = $request->input('idItem') ;
+        $estoque_cadastrar->idFornecedor = $request->input('idFornecedor') ;
+        $estoque_cadastrar->idLocalidade = $request->input('idLocalidade') ;
+        $estoque_cadastrar->estoque_min = $request->input('estoque_min') ;
+        $estoque_cadastrar->data_entrada = now();
+
+        //SALVADO NO BANCO DE DADOS NA TABELA "ESTOQUE"
+        $estoque_cadastrar->save();
+
+        //APÓS A INSERSÃO ELE RETORNA PRA TELA DE LISTAGEM
+        return redirect()->route('listar_estoque')
 
         ->with(compact('estoque'))
 
@@ -97,9 +117,16 @@ class CadastrarEstoqueController extends Controller
 
     public function listar(Request $request){
 
+        if(!\Gate::allows('isAdmin')){
+            abort(403,"Não pode executar essa ação");
+        }
+
+        //INSTANCIANDO A MODEL DE ESTOQUE NA VARIÁVEL $estoque
         $estoque = new Estoque();
 
+        //USANDO A FUNÇÃO "SEARCH" PARA O CAMPO "PESQUISAR" NA TELA DE LISTAGEM
         $search = $request->get('search');
+        $search_localidade = $request->get('search_localidade');
                 
         $item = new Item();
         $item = $item::all();
@@ -110,15 +137,18 @@ class CadastrarEstoqueController extends Controller
         $localidade = new Localidade();
         $localidade = $localidade::all();
 
+        //CONSULTA TRAZENDO TODOS OS DADOS INSERIDOS NA TABELA DE ESTOQUE, COM A OPÇÃO DE PESQUISA USANDO A FUNÇÃO "SEARCH" NO WHERE
         $estoque = DB::table ('tbEstoque')
         ->join('tbItem', 'tbEstoque.idItem', '=', 'tbItem.idItem')
         ->join('tbFornecedor', 'tbEstoque.idFornecedor', '=', 'tbFornecedor.idFornecedor')
         ->join('tbLocalidade', 'tbEstoque.idLocalidade', '=', 'tbLocalidade.idLocalidade')
-        ->select('idEstoque','quantidade_total','data_entrada','numero_nf','data_nf','data_garantia','tbItem.descricao_item', 'tbFornecedor.razao_social', 'tbLocalidade.localidade', 'tbLocalidade.setor') 
+        ->select('idEstoque','quantidade_total','data_entrada','numero_nf','data_nf','data_garantia','tbItem.descricao_item', 'tbFornecedor.razao_social', 'tbLocalidade.localidade', 'tbLocalidade.setor', 'estoque_min') 
         ->where('tbItem.descricao_item', 'LIKE', '%'.$search.'%')
+        ->where('tbLocalidade.localidade', 'LIKE', '%'.$search_localidade.'%')
         ->Paginate(10);
 
 
+        //RETORNANDO PARA TELA DE LISTAGEM
         return view('listar_estoque')
 
         ->with(compact('estoque'))
@@ -129,20 +159,26 @@ class CadastrarEstoqueController extends Controller
 
         ->with(compact('search'))
 
+        ->with(compact('search_localidade'))
+
         ->with(compact('fornecedor'));
         
     }
 
     public function edit($id)
     {
-       
+        if(!\Gate::allows('isAdmin')){
+            abort(403,"Não pode executar essa ação");
+        }
+        //O EDIT É UM SELECT COM UM WHERE PUXANDO O ID DO ITEM CADASTRADO/SELECIONADO
         $editar = new Estoque();
 
+        //CONSULTA COM O WHERE idEstoque = $id 
         $editar = DB::table ('tbEstoque')
         ->join('tbItem', 'tbEstoque.idItem', '=', 'tbItem.idItem')
         ->join('tbFornecedor', 'tbEstoque.idFornecedor', '=', 'tbFornecedor.idFornecedor')
         ->join('tbLocalidade', 'tbEstoque.idLocalidade', '=', 'tbLocalidade.idLocalidade')
-        ->select('idEstoque','quantidade_total','data_entrada','numero_nf','data_nf','data_garantia','tbItem.descricao_item', 'tbItem.idItem',  'tbFornecedor.razao_social', 'tbFornecedor.idFornecedor', 'tbLocalidade.localidade', 'tbLocalidade.setor') 
+        ->select('idEstoque','quantidade_total','data_entrada','numero_nf','data_nf','data_garantia','tbItem.descricao_item', 'tbItem.idItem',  'tbFornecedor.razao_social', 'tbFornecedor.idFornecedor', 'tbLocalidade.localidade', 'tbLocalidade.setor', 'tbLocalidade.idLocalidade','estoque_min' ) 
         ->where('idEstoque', $id)
         ->first();
 
@@ -157,7 +193,7 @@ class CadastrarEstoqueController extends Controller
         $localidade = $localidade::all();
         
 
-       
+        //INDO PRA TELA DE EDIÇÃO PARA DAR UM UPDATE NAS INFORMAÇÕES
         return view('editar_estoque')
 
         ->with(compact('editar'))
@@ -171,7 +207,11 @@ class CadastrarEstoqueController extends Controller
     }
 
     public function update(Request $request, $id)
-    {
+    {   
+        if(!\Gate::allows('isAdmin')){
+            abort(403,"Não pode executar essa ação");
+        }
+        //UPDATE NOS DADOS MODIFICADOS (OU NÃO)  
         $editar = new Estoque();
         $editar = DB::table ('tbEstoque')
         ->where('idEstoque', $id)
@@ -182,10 +222,11 @@ class CadastrarEstoqueController extends Controller
         'data_garantia' =>$request->input('data_garantia'),
         'idItem' =>$request->input('idItem'),
         'idFornecedor' =>$request->input('idFornecedor'),
+        'estoque_min' =>$request->input('estoque_min'),
         'idLocalidade' =>$request->input('idLocalidade')]
         );
        
-            
+        //SE EXISTIR O UPDATE ELE VOLTA PRA TELA DE HOME
         if(isset($editar)){
             
         return redirect()->route('home');
@@ -195,15 +236,19 @@ class CadastrarEstoqueController extends Controller
 
     public function destroy($id)
     {
+        if(!\Gate::allows('isAdmin')){
+            abort(403,"Não pode executar essa ação");
+        }
         $estoque = new Estoque();
         
-
+        //DELETE NO ESTOQUE SELECIONADO
         $deletedRows =  $estoque::where('idEstoque', $id)->delete();
         
-        
+               
+        return redirect()->route('home')
 
-        if(isset($estoque)){
-            return redirect()->route('home', compact('estoque'));
-        }
+        ->with(compact('estoque'));
+
+
     }
 }
